Refined code.
The app produces sensor data consisting of the nodeId, timeStamp and randomly generated 5 bytes.
The client reads sensor data from the server. With each sensor data element that is read, it is deleted on the server side. The client continues reading till the server data list is empty. The client orients the server and disconnects.

Writing to file at the sink is remaining.